==> 1_load_data.ipynb <==
### If you need to install matplotlib uncomment the following line and run this code block.
# !pip install matplotlib

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')

variants = pd.read_csv("Input_data/variant_cases.csv", sep=",", index_col=0)
vaccinations = pd.read_csv("Input_data/vaccinations.csv", sep=',', index_col=0)

def drop_data(row):
    date = row["year_week"].split("-")
    if int(date[0]) == 2020:
        if int(date[1]) < 40:
            return "drop"

variants = variants.loc[(variants['variant'] == 'B.1.617.2') | (variants['variant'] == 'B.1.1.529') |\
                       (variants['variant'] == 'B.1.1.7') | (variants['variant'] == 'B.1.351')]
variants["drop_column"] = variants.apply(drop_data, axis = 1)
variants = variants.loc[variants["drop_column"] != "drop"]
variants = variants.drop(columns = "drop_column")
variants

sort_variants = variants.sort_values('number_sequenced')
var = sort_variants.drop_duplicates(subset=['country_code', 'year_week', 'variant'], keep='last')
variants = var.sort_values(by=['country_code','year_week'])
variants

def get_variants(row):
    percentage = row['percent_variant']/100
    cases = row['new_cases'] * percentage
    return cases
    
variants['number_cases_ratio'] = variants.apply(get_variants, axis=1)
variants = variants.drop(columns=['source', 'number_sequenced', 'percent_cases_sequenced', 'valid_denominator',\
                                  'number_detections_variant', 'number_sequenced_known_variant'])
variants

variants = variants.loc[(variants['variant'] != 'B.1.351')]
variants

vaccinations = vaccinations.loc[vaccinations['TargetGroup'] == 'ALL']
vaccinations = vaccinations.drop_duplicates(subset='ReportingCountry')
vaccinations

def get_population(row, df):
    pop = df['Population'].loc[df['ReportingCountry'] == row['country_code']]
    return pop[0]

variants['population'] = variants.apply(lambda row: get_population(row,vaccinations), axis=1)
variants

def normalize_counts(row):
    percentage = (row['number_cases_ratio']/row['population']) * 100000
    return percentage
    
variants['normalized_counts'] = variants.apply(normalize_counts, axis=1)

variants = variants.reset_index()
variants

variant_data = variants[["variant","normalized_counts"]]#.reset_index(drop=True)
variant_data_p = variant_data.pivot(columns = ["variant"], values = ["normalized_counts"])
variant_data_p = variant_data_p.droplevel(level = 0, axis = 1)
week_data = pd.DataFrame(variants[["country","country_code","year_week","new_cases","population"]]).reset_index(drop=True)
week_data.index.rename("variant", inplace = True)
combined = week_data.join(variant_data_p, on="variant").reset_index(drop=True)
combined

data = []
for idx, row in combined.iterrows():
    row_list = []
    if idx > 1:
        prev = (idx - 1)
        if row["year_week"] == list(combined.loc[combined.index == prev]["year_week"])[0]:
            if pd.isna(row["B.1.1.529"]):
                row["B.1.1.529"] = list(combined.loc[combined.index == prev]["B.1.1.529"])[0]
            if pd.isna(row["B.1.1.7"]):
                row["B.1.1.7"] = list(combined.loc[combined.index == prev]["B.1.1.7"])[0]
            if pd.isna(row["B.1.617.2"]):
                row["B.1.617.2"] = list(combined.loc[combined.index == prev]["B.1.617.2"])[0]
        if row["year_week"] == list(combined.loc[combined.index == prev - 1]["year_week"])[0]:
            if pd.isna(row["B.1.1.529"]):
                row["B.1.1.529"] = list(combined.loc[combined.index == prev - 1]["B.1.1.529"])[0]
            if pd.isna(row["B.1.1.7"]):
                row["B.1.1.7"] = list(combined.loc[combined.index == prev - 1]["B.1.1.7"])[0]
            if pd.isna(row["B.1.617.2"]):
                row["B.1.617.2"] = list(combined.loc[combined.index == prev - 1]["B.1.617.2"])[0]
                
        row_list.append(row["country"]) 
        row_list.append(row["country_code"])
        row_list.append(row["year_week"])
        row_list.append(row["new_cases"])
        row_list.append(row["B.1.1.7"])
        row_list.append(row["B.1.617.2"])
        row_list.append(row["B.1.1.529"])
        row_list.append(row["population"])
        data.append(row_list)

updated_df = pd.DataFrame(data,columns=["Country","Country_Code","Year_Week","New_Cases","Alpha","Delta","Omicron",
                                       "Population"])
updated_df = updated_df.dropna(thresh=6)
updated_df = updated_df.drop_duplicates(subset=['Country_Code', 'Year_Week'], keep='last')
updated_df = updated_df.set_index("Country")
updated_df = updated_df.fillna(0)
updated_df

updated_df.to_csv("tidy_variant_data.tsv", sep="\t")
