==> 3_make_gif.ipynb <==
from PIL import Image
import glob
 
# Create the frames
frames = []
imgs = sorted(glob.glob("png_files/*.png"))
for i in imgs:
    new_frame = Image.open(i)
    frames.append(new_frame)
 
# Save into a GIF file that loops forever
frames[0].save('web-setup/front-end/public/map1.gif', format='GIF',
               append_images=frames[1:],
               save_all=True,
               duration=500, loop=0)
